IfStatements completed and working
May be a few kinks here and there, but should be pretty much done. Also needs else, and the ability to omit braces (I always do that, and it is currently throwing me off). Also next may be a while loop, as well as comparators like less than or greater than.